backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.2.0
